% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detatch_tables.R
\name{detatch_database}
\alias{detatch_database}
\title{Detach a database namespace and close the connection}
\usage{
detatch_database(dbname)
}
\arguments{
\item{dbname}{}
}
\value{
TRUE invisibly if detach was successful
}
\description{
Detach a database namespace and close the connection
}
