% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_tables.R
\name{attach_database}
\alias{attach_database}
\title{Attach all the tables in a database as tbl objects within an attached environment}
\usage{
attach_database(db, dbname)
}
\arguments{
\item{db}{A \link[DBI:dbConnection]{DBI::dbConnection} object}

\item{dbname}{Character. Name of the environment to create}
}
\value{
Invisibly returns character vector of table names.
}
\description{
Attach all the tables in a database as tbl objects within an attached environment
}
